requirements.txt
Flask==0.12
mongoengine==0.11.0
pkg-resources==0.0.0
pycrypto==2.6.1
quickconfig==2.1
requests==2.13.0
